move table reflection to deprecated namespace
